add int64 support distinct from int

+++ bytecode/op_compare.go
@@ -82,7 +82,7 @@ func equalByteCode(c *Context, i interface{}) *errors.EgoError {
 				r = false
 
 			case byte, int32, int, int64:
-				r = datatypes.GetInt(v1) == datatypes.GetInt(v2)
+				r = datatypes.GetInt64(v1) == datatypes.GetInt64(v2)
 
 			case float64:
 				r = v1.(float64) == v2.(float64)
@@ -165,7 +165,7 @@ func notEqualByteCode(c *Context, i interface{}) *errors.EgoError {
 			r = false
 
 		case byte, int32, int, int64:
-			r = datatypes.GetInt(v1) != datatypes.GetInt(v2)
+			r = datatypes.GetInt64(v1) != datatypes.GetInt64(v2)
 
 		case float32:
 			r = v1.(float32) != v2.(float32)
@@ -226,7 +226,7 @@ func greaterThanByteCode(c *Context, i interface{}) *errors.EgoError {
 
 		switch v1.(type) {
 		case byte, int32, int, int64:
-			r = datatypes.GetInt(v1) > datatypes.GetInt(v2)
+			r = datatypes.GetInt64(v1) > datatypes.GetInt64(v2)
 
 		case float32:
 			r = v1.(float32) > v2.(float32)
@@ -288,7 +288,7 @@ func greaterThanOrEqualByteCode(c *Context, i interface{}) *errors.EgoError {
 
 		switch v1.(type) {
 		case byte, int32, int, int64:
-			r = datatypes.GetInt(v1) >= datatypes.GetInt(v2)
+			r = datatypes.GetInt64(v1) >= datatypes.GetInt64(v2)
 
 		case float32:
 			r = v1.(float32) >= v2.(float32)
@@ -351,7 +351,7 @@ func lessThanByteCode(c *Context, i interface{}) *errors.EgoError {
 
 		switch v1.(type) {
 		case byte, int32, int, int64:
-			r = datatypes.GetInt(v1) < datatypes.GetInt(v2)
+			r = datatypes.GetInt64(v1) < datatypes.GetInt64(v2)
 
 		case float32:
 			r = v1.(float32) < v2.(float32)
@@ -412,7 +412,7 @@ func lessThanOrEqualByteCode(c *Context, i interface{}) *errors.EgoError {
 		v1, v2 = datatypes.Normalize(v1, v2)
 		switch v1.(type) {
 		case byte, int32, int, int64:
-			r = datatypes.GetInt(v1) <= datatypes.GetInt(v2)
+			r = datatypes.GetInt64(v1) <= datatypes.GetInt64(v2)
 
 		case float32:
 			r = v1.(float32) <= v2.(float32)
